added dataxy(array) in encoding
dataxy array for checking collision

+++ Utils/MaxDataCalculator.py
@@ -2,4 +2,4 @@
 #returning max number of char that can be hide into img
 def MaxASCIIchar(img,thresold = 0.3):
 	h , w , c = img.shape
-	return h*w*thresold
+	return int(h*w*thresold)
